RavenDB-4345 Remove LazyLoadOperation and use LazyMultiLoadOperation as that this is the expected result from the server in v4.0. We rename the LazyMultiLoadOperation to be LazyLoadOperation
