+++ src/Raven.Client/Connection/Async/IAsyncDatabaseCommands.cs
@@ -148,7 +148,7 @@ namespace Raven.Client.Connection.Async
         /// <param name="transformerParameters">parameters that will be passed to transformer</param>
         /// <param name="metadataOnly">specifies if only document metadata should be returned</param>
         /// <param name="token">The cancellation token.</param>
-        Task<MultiLoadResult> GetAsync(string[] keys, string[] includes, string transformer = null, Dictionary<string, RavenJToken> transformerParameters = null, bool metadataOnly = false, CancellationToken token = default (CancellationToken));
+        Task<LoadResult> GetAsync(string[] keys, string[] includes, string transformer = null, Dictionary<string, RavenJToken> transformerParameters = null, bool metadataOnly = false, CancellationToken token = default (CancellationToken));
 
         /// <summary>
         ///     Get the low level bulk insert operation
@@ -333,7 +333,7 @@ namespace Raven.Client.Connection.Async
         /// </summary>
         /// <param name="query">more like this query definition that will be executed</param>
         /// <param name="token">The cancellation token.</param>
-        Task<MultiLoadResult> MoreLikeThisAsync(MoreLikeThisQuery query, CancellationToken token = default(CancellationToken));
+        Task<LoadResult> MoreLikeThisAsync(MoreLikeThisQuery query, CancellationToken token = default(CancellationToken));
 
         /// <summary>
         ///     Perform a single POST request containing multiple nested GET requests
