+++ src/Raven.Client/Data/LoadResult.cs
@@ -1,5 +1,5 @@
 //-----------------------------------------------------------------------
-// <copyright file="MultiLoadResult.cs" company="Hibernating Rhinos LTD">
+// <copyright file="LoadResult.cs" company="Hibernating Rhinos LTD">
 //     Copyright (c) Hibernating Rhinos LTD. All rights reserved.
 // </copyright>
 //-----------------------------------------------------------------------
@@ -11,7 +11,7 @@ namespace Raven.Abstractions.Data
     /// <summary>
     /// Represent a result which include both document results and included documents
     /// </summary>
-    public class MultiLoadResult
+    public class LoadResult
     {
         /// <summary>
         /// Loaded documents. The results will be in exact same order as in keys parameter.
@@ -24,9 +24,9 @@ namespace Raven.Abstractions.Data
         public List<RavenJObject> Includes { get; set; }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="MultiLoadResult"/> class.
+        /// Initializes a new instance of the <see cref="LoadResult"/> class.
         /// </summary>
-        public MultiLoadResult()
+        public LoadResult()
         {
             Results = new List<RavenJObject>();
             Includes = new List<RavenJObject>();
